+++ src/main/java/org/example/repository/AstartesSpecialRangedWeaponRepository.java
@@ -4,7 +4,7 @@ import java.util.List;
 public interface AstartesSpecialRangedWeaponRepository {
 
 
-    AstartesSpecialRangedWeaponEntity get(Long id);
+    AstartesSpecialRangedWeaponEntity get(int id);
     void add(AstartesSpecialRangedWeaponEntity astartesSpecialRangedWeaponEntity);
     void update(AstartesSpecialRangedWeaponEntity astartesSpecialRangedWeaponEntity);
     void remove(AstartesSpecialRangedWeaponEntity astartesSpecialRangedWeaponEntity);
